Form finally working with axios. Validation and styling all done with Formik.
