+++ package.json
@@ -5,6 +5,7 @@
   "version": "0.1.0",
   "author": "Kyle Mathews <mathews.kyle@gmail.com>",
   "dependencies": {
+    "axios": "^0.19.2",
     "env-cmd": "^10.1.0",
     "formik": "^2.1.4",
     "gatsby": "^2.19.7",
